made it so create_file does not attempt to create a file that already exists.

+++ src/commands/utils.rs
@@ -8,13 +8,19 @@ pub fn create_file(name: &String, extension: &String, contents: &String) {
     let file_path: &Path = Path::new(filename);
     let display = file_path.display();
 
-    let mut file = match File::create(&file_path) {
-        Err(e) => panic!("couldn't create {}: {}", display, e),
-        Ok(file) => file,
-    };
+    if file_path.is_file() {
+        println!("Skipping creating {} as it already exists!", display);
+    } else {
+        let mut file = match File::create(&file_path) {
+            Err(e) => panic!("couldn't create {}: {}", display, e),
+            Ok(file) => file,
+        };
 
-    match file.write_all(contents.as_bytes()) {
-        Err(e) => panic!("Write failed at {} because of {}.", display, e),
-        Ok(_) => println!("{} successfully created!", display),
+        match file.write_all(contents.as_bytes()) {
+            Err(e) => panic!("Write failed at {} because of {}.", display, e),
+            Ok(_) => println!("{} successfully created!", display),
+        }
     }
+
+
 }
